+++ web/src/Api/index.js
@@ -243,3 +243,21 @@ export const spotifyDisonnectionRequest = () => {
   });
 }
 
+
+export const getServerInfo = () => {
+  return axios.get(process.env.REACT_APP_API_URL + "serverinfo")
+  .then((data) => {
+    return data.data
+  }).catch((error) => {
+    return handleErrorResp(error)
+  });
+}
+
+export const resetScrobbleToken = () => {
+  return axios.patch(process.env.REACT_APP_API_URL + "user", { token: "" }, { headers: getHeaders() })
+  .then((data) => {
+    return data.data
+  }).catch((error) => {
+    return handleErrorResp(error)
+  });
+}